@@ -13,38 +13,37 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@angular/animations": "~12.2.13",
-    "@angular/cdk": "^12.2.12",
-    "@angular/common": "~12.2.13",
-    "@angular/compiler": "~12.2.13",
-    "@angular/core": "~12.2.13",
+    "@angular/animations": "~13.0.0",
+    "@angular/cdk": "^13.0.0",
+    "@angular/common": "~13.0.0",
+    "@angular/compiler": "~13.0.0",
+    "@angular/core": "~13.0.0",
     "@angular/fire": "^7.1.1",
-    "@angular/forms": "~12.2.13",
-    "@angular/material": "^12.2.12",
-    "@angular/platform-browser": "~12.2.13",
-    "@angular/platform-browser-dynamic": "~12.2.13",
-    "@angular/router": "~12.2.13",
+    "@angular/forms": "~13.0.0",
+    "@angular/material": "^13.0.0",
+    "@angular/platform-browser": "~13.0.0",
+    "@angular/platform-browser-dynamic": "~13.0.0",
+    "@angular/router": "~13.0.0",
     "@swimlane/ngx-charts": "^19.1.0",
     "angular-responsive-carousel": "^2.1.2",
     "firebase": "^9.3.0",
     "get-video-id": "^3.4.3",
     "google-libphonenumber": "^3.2.25",
     "rxfire": "^6.0.3",
-    "rxjs": "~6.6.7",
+    "rxjs": "~7.4.0",
     "tslib": "^2.3.1",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1202.13",
-    "@angular-devkit/build-angular": "~12.2.13",
+    "@angular-devkit/architect": "^0.1300.1",
+    "@angular-devkit/build-angular": "~13.0.1",
     "@angular-eslint/builder": "12.6.1",
     "@angular-eslint/eslint-plugin": "12.6.1",
     "@angular-eslint/eslint-plugin-template": "12.6.1",
     "@angular-eslint/schematics": "12.6.1",
     "@angular-eslint/template-parser": "12.6.1",
-    "@angular/cli": "~12.2.13",
-    "@angular/compiler-cli": "~12.2.13",
-    "@netlify-builder/deploy": "^4.1.2",
+    "@angular/cli": "~13.0.1",
+    "@angular/compiler-cli": "~13.0.0",
     "@types/d3-scale": "^4.0.2",
     "@types/d3-selection": "^3.0.1",
     "@types/google-libphonenumber": "^7.4.22",
@@ -61,7 +60,7 @@
     "inquirer-autocomplete-prompt": "^1.4.0",
     "jasmine-core": "~3.10.1",
     "jsonc-parser": "^3.0.0",
-    "karma": "~6.3.7",
+    "karma": "~6.3.8",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.1",
@@ -74,7 +73,7 @@
     "stylelint-config-property-sort-order-smacss": "^8.0.0",
     "stylelint-config-standard": "^23.0.0",
     "stylelint-order": "^5.0.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.4.4"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx}": "eslint --cache --fix",
