@@ -2,7 +2,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { Injectable } from '@angular/core'
 import { Router } from '@angular/router'
-import { BehaviorSubject, zip } from 'rxjs'
+import { BehaviorSubject, lastValueFrom, zip } from 'rxjs'
 import { map, take } from 'rxjs/operators'
 import { MatSnackBar } from '@angular/material/snack-bar'
 import { HttpClient } from '@angular/common/http'
@@ -30,6 +30,7 @@ import {
   CreateUserRequest,
   CreateUserResponse,
   GetUserResponse,
+  GetUserIdResponse,
 } from 'src/app/shared/models/api.interface'
 
 import { UpdateAccountDetailsForm } from '../../containers/update-account-details/update-account-details.component'
@@ -112,7 +113,7 @@ export class AuthService {
       this._displayName = credential.user.displayName
 
       // User Registration
-      const { claims, uid } = await this.user$.pipe(take(1)).toPromise()
+      const { claims, uid } = await lastValueFrom(this.user$.pipe(take(1)))
       isNewUser = getAdditionalUserInfo(credential)?.isNewUser ?? false
       if (isNewUser) {
         // Permissions
@@ -128,6 +129,7 @@ export class AuthService {
       if (!isExtraDataComplete) return void this.router.navigate(['/auth/extra-data'])
 
       // Completed registration operations
+      void this.setUserId()
       await this.router.navigate(['/'])
       return void this._verifyEmail()
     } catch (error: any) {
@@ -142,10 +144,11 @@ export class AuthService {
     try {
       if (!token) return false
 
-      return this.http
-        .get<never>(`${environment.apiUrl}/auth/${token}`)
-        .pipe(map(({ body }) => !!body && body !== 'Token is invalid'))
-        .toPromise()
+      return lastValueFrom(
+        this.http
+          .get<never>(`${environment.apiUrl}/auth/${token}`)
+          .pipe(map(({ body }) => !!body && body !== 'Token is invalid')),
+      )
     } catch (error: any) {
       error.code
         ? this.errorHandler.openDialog(error.code as string)
@@ -157,7 +160,7 @@ export class AuthService {
 
   private async _verifyEmail(): Promise<void> {
     try {
-      const { isEmailVerified } = await this.user$.pipe(take(1)).toPromise()
+      const { isEmailVerified } = await lastValueFrom(this.user$.pipe(take(1)))
 
       if (isEmailVerified || !this._user) return
 
@@ -183,9 +186,9 @@ export class AuthService {
 
   private async _verifyExtraDataCompleted(uid: User['uid']): Promise<boolean> {
     try {
-      const response = await this.http
-        .get<GetUserResponse>(`${environment.apiUrl}/users/${uid}`)
-        .toPromise()
+      const response = await lastValueFrom(
+        this.http.get<GetUserResponse>(`${environment.apiUrl}/users/${uid}`),
+      )
 
       if (!response.email) {
         await this.logout()
@@ -261,9 +264,9 @@ export class AuthService {
         phoneNumber: null,
       }
 
-      const response = await this.http
-        .post<CreateUserResponse>(`${environment.apiUrl}/users/?uid=${uid}`, body)
-        .toPromise()
+      const response = await lastValueFrom(
+        this.http.post<CreateUserResponse>(`${environment.apiUrl}/users/?uid=${uid}`, body),
+      )
 
       return !!response?.id
     } catch (error: any) {
@@ -276,15 +279,15 @@ export class AuthService {
   }
 
   private async _sendUserData(uid: User['uid'], body: CreateUserRequest): Promise<boolean> {
-    const response = await this.http
-      .put(
+    const response = await lastValueFrom(
+      this.http.put(
         `${environment.apiUrl}/users/?uid=${uid}`,
         { ...body, province: await this.location.getIdByProvince(body.province) },
         {
           responseType: 'text',
         },
-      )
-      .toPromise()
+      ),
+    )
 
     return response === 'Usuario actualizado'
   }
@@ -293,7 +296,7 @@ export class AuthService {
     try {
       const { fullName, birthday, province, locality, phoneNumber } = data
 
-      const { uid } = await this.user$.pipe(take(1)).toPromise()
+      const { uid } = await lastValueFrom(this.user$.pipe(take(1)))
 
       const body = {
         displayName: fullName,
@@ -303,11 +306,12 @@ export class AuthService {
         phoneNumber: convertPhoneNumber(phoneNumber),
       }
 
-      await this.http
-        .put(`${environment.apiUrl}/users/?uid=${uid}`, body, {
+      await lastValueFrom(
+        this.http.put(`${environment.apiUrl}/users/?uid=${uid}`, body, {
           responseType: 'text',
-        })
-        .toPromise()
+        }),
+      )
+
       await this._updateUser()
     } catch (error: any) {
       error.code
@@ -322,7 +326,7 @@ export class AuthService {
     try {
       const { birthday, phoneNumber, province, locality } = extraData
 
-      const { uid, displayName } = await this.user$.pipe(take(1)).toPromise()
+      const { uid, displayName } = await lastValueFrom(this.user$.pipe(take(1)))
 
       await this._sendUserData(uid, {
         displayName,
@@ -350,9 +354,9 @@ export class AuthService {
   async setPermissions(type: 'donor' | 'admin', uid: User['uid']): Promise<void | boolean> {
     try {
       const body = type === 'admin' ? { admin: true, donor: false } : { admin: false, donor: true }
-      await this.http
-        .put<User['claims']>(`${environment.apiUrl}/users/claims/${uid}`, body)
-        .toPromise()
+      await lastValueFrom(
+        this.http.put<User['claims']>(`${environment.apiUrl}/users/claims/${uid}`, body),
+      )
 
       void this._updateUser()
 
@@ -378,12 +382,13 @@ export class AuthService {
   }
 
   async verifyAdminClaim(): Promise<boolean> {
-    const { uid } = await this.user$.pipe(take(1)).toPromise()
+    const { uid } = await lastValueFrom(this.user$.pipe(take(1)))
 
-    const response = await this.http
-      .get<never>(`${environment.apiUrl}/auth/claims?uid=${uid}`)
-      .pipe(map(({ body }) => body))
-      .toPromise()
+    const response = await lastValueFrom(
+      this.http
+        .get<never>(`${environment.apiUrl}/auth/claims?uid=${uid}`)
+        .pipe(map(({ body }) => body)),
+    )
 
     return response
   }
@@ -411,6 +416,7 @@ export class AuthService {
       await this.auth.signOut()
       localStorage.removeItem('token')
       localStorage.removeItem('claims')
+      localStorage.removeItem('userId')
       this._displayName = null
       this.snackBar.open('Se ha cerrado tu sesión correctamente', 'Cerrar', { duration: 5000 })
       return void this.router.navigate(['/'])
@@ -440,16 +446,17 @@ export class AuthService {
   }
 
   async deleteUser(): Promise<void> {
-    const isApproved = (await this.dialog
-      .open<DialogComponent, DialogData>(DialogComponent, {
-        data: {
-          actions: ['No', 'Sí, eliminar'],
-          title: '¿Estás seguro de eliminar tu cuenta?',
-          description: 'La eliminación es definitiva. No podrás volver a reactivar la cuenta.',
-        },
-      })
-      .afterClosed()
-      .toPromise()) as boolean
+    const isApproved = (await lastValueFrom(
+      this.dialog
+        .open<DialogComponent, DialogData>(DialogComponent, {
+          data: {
+            actions: ['No', 'Sí, eliminar'],
+            title: '¿Estás seguro de eliminar tu cuenta?',
+            description: 'La eliminación es definitiva. No podrás volver a reactivar la cuenta.',
+          },
+        })
+        .afterClosed(),
+    )) as boolean
     if (!isApproved) return
 
     await this.auth.signOut()
@@ -464,5 +471,13 @@ export class AuthService {
       const token = await getIdToken(this._user, true)
       localStorage.setItem('token', token)
     }
+  }
+
+  async setUserId(): Promise<void> {
+    const uid = (await lastValueFrom(this.user$.pipe(take(1)))).uid
+    const { id } = await lastValueFrom(
+      this.http.get<GetUserIdResponse>(`${environment.apiUrl}/users/${uid}`),
+    )
+    localStorage.setItem('userId', id.toString())
   }
 }
