@@ -1,19 +1,23 @@
 import { animate, state, style, transition, trigger } from '@angular/animations'
-import { Component } from '@angular/core'
-import {
-  DonationsService,
-  DonationTest,
-} from 'src/app/donation/services/donations/donations.service'
+import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core'
+import { ActivatedRoute } from '@angular/router'
+import { DonationsService } from 'src/app/donation/services/donations/donations.service'
+import { GetDonorsByFiltersResponse } from 'src/app/shared/models/api.interface'
+import { DonationWithProjectName } from '../../../donation/services/donations/donations.service'
+import { FiltersData, OrdersData } from '../../containers/project-metrics/project-metrics.component'
+import { GetDonorsByFiltersArgs, MetricsService } from '../../services/metrics/metrics.service'
 
 interface DonorData {
+  id: number
   fullName: string
   province: string
   locality: string
   email: string
+  photo: string | null
   number: string
   donationsQuantity: number
   totalAmount: number
-  donations: DonationTest[]
+  donations?: DonationWithProjectName[]
 }
 
 @Component({
@@ -28,73 +32,147 @@ interface DonorData {
     ]),
   ],
 })
-export class MetricsListComponent {
+export class MetricsListComponent implements OnInit, OnChanges {
+  // Inputs
+  @Input() filtersData!: FiltersData
+  @Input() ordersData!: OrdersData
+
+  // General settings
   columnNames = {
     fullName: 'Donante',
     donationsQuantity: 'Cantidad de Donaciones',
     totalAmount: 'Importe donado',
     action: ' ',
   }
-
-  dataSource: DonorData[] = [
-    {
-      fullName: 'Rodrigo Moyano',
-      province: 'San Juan',
-      locality: 'Capital',
-      email: 'email@email.com',
-      number: '265 587 4332',
-      donationsQuantity: 5,
-      totalAmount: 730,
-      donations: this.donations.getDonations(),
-    },
-    {
-      fullName: 'Juan José',
-      province: 'San Juan',
-      locality: 'Capital',
-      email: 'email@email.com',
-      number: '265 587 4332',
-      donationsQuantity: 234,
-      totalAmount: 4540,
-      donations: this.donations.getDonations(),
-    },
-    {
-      fullName: 'Francisco González',
-      province: 'San Juan',
-      locality: 'Capital',
-      email: 'email@email.com',
-      number: '265 587 4332',
-      donationsQuantity: 4,
-      totalAmount: 980,
-      donations: this.donations.getDonations(),
-    },
-    {
-      fullName: 'Macarena Pérez',
-      province: 'San Juan',
-      locality: 'Capital',
-      email: 'email@email.com',
-      number: '265 587 4332',
-      donationsQuantity: 53,
-      totalAmount: 430,
-      donations: this.donations.getDonations(),
-    },
-    {
-      fullName: 'Sherlock Holmes',
-      province: 'San Juan',
-      locality: 'Capital',
-      email: 'email@email.com',
-      number: '265 587 4332',
-      donationsQuantity: 44,
-      totalAmount: 7530,
-      donations: this.donations.getDonations(),
-    },
-  ]
   columnsToDisplay: ['fullName', 'donationsQuantity', 'totalAmount', 'action'] = [
     'fullName',
     'donationsQuantity',
     'totalAmount',
     'action',
   ]
-  expandedElement: DonorData | null = null
 
-  constructor(private donations: DonationsService) {}
+  // Data
+  donorsData!: DonorData[]
+  metricsData!: GetDonorsByFiltersResponse['body']
+  selectedProjectId: string | null = this.route.snapshot.paramMap.get('id')
+
+  // States
+  expandedElement: DonorData | null = null
+  constructor(
+    private donations: DonationsService,
+    private metrics: MetricsService,
+    private route: ActivatedRoute,
+  ) {}
+
+  async ngOnInit(): Promise<void> {
+    await this.getDonorsData()
+    this.convertDonorsData()
+  }
+
+  async ngOnChanges(): Promise<void> {
+    await this.getDonorsData()
+    this.convertDonorsData()
+  }
+
+  async getDonorsData(): Promise<void> {
+    const projectId = Number(this.selectedProjectId)
+
+    const selectedOrder: GetDonorsByFiltersArgs = this.ordersData
+      ? Object.fromEntries([[this.ordersData.type, this.ordersData.value]])
+      : {
+          orderAlphabetically: 'ascending',
+        }
+
+    if (!this.filtersData) {
+      this.metricsData = await this.metrics.getDonorsByFilters({
+        projectId,
+
+        ...selectedOrder,
+      })
+
+      return
+    }
+
+    const { province, age1, age2, amount1, amount2, paymentType: isRecurring } = this.filtersData
+
+    this.metricsData = await this.metrics.getDonorsByFilters({
+      // General
+      projectId,
+
+      page: 0, // Current Page
+      size: 30, // Items per page
+
+      // Filters
+      provinceId: province ?? undefined,
+
+      age1: Number(age1) ?? undefined,
+      age2: Number(age2) ?? undefined,
+
+      mount1: Number(amount1) ?? undefined,
+      mount2: Number(amount2) ?? undefined,
+
+      type: isRecurring ? 'recurring' : 'regular',
+
+      // Orders
+      ...selectedOrder,
+    })
+  }
+
+  convertDonorsData(): void {
+    if (!this.metricsData) return
+
+    const { data: donors } = this.metricsData
+
+    this.donorsData = donors.map(
+      ({ donationCount: donationsQuantity, amount: totalAmount, user }) => {
+        const {
+          id,
+          displayName: fullName,
+          province,
+          locality,
+          email,
+          phoneNumber: number,
+          photoURL: photo,
+        } = user
+
+        return {
+          id,
+          fullName,
+          province: province.name,
+          locality,
+          email,
+          number,
+          photo,
+          donationsQuantity,
+          totalAmount,
+        }
+      },
+    )
+  }
+
+  async selectStateOfExpandedElement(element: DonorData): Promise<void> {
+    this.expandedElement = this.expandedElement === element ? null : element
+
+    this.donorsData = await Promise.all(
+      this.donorsData.map(async (data) => {
+        if (data === this.expandedElement) {
+          const donationsByUser = await this.donations.getDonationsByUserAndProject(
+            data.id,
+            Number(this.selectedProjectId),
+          )
+
+          const modifiedData = {
+            ...data,
+            donations: donationsByUser,
+          }
+
+          this.expandedElement = modifiedData
+
+          return modifiedData
+        }
+
+        return data
+      }),
+    )
+  }
 }
