@@ -1,8 +1,11 @@
 import { Component } from '@angular/core'
+import { AuthService } from 'src/app/auth/services/auth/auth.service'
 
 @Component({
   selector: 'app-success-payment',
   templateUrl: './success-payment.component.html',
   styleUrls: ['./success-payment.component.css'],
 })
-export class SuccessPaymentComponent {}
+export class SuccessPaymentComponent {
+  constructor(public auth: AuthService) {}
+}
